S02P23B204-56, S02P23B204-59|Update Entity Columns Modified

See merge request s02-bigdata-sub3/s02p23b204!23

+++ back/findme/src/main/java/com/ssafy/findme/repository/UnsympRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.findme.domain.Unsymp;
 
-public interface UnsympRepository extends JpaRepository<Unsymp, Integer> {
+public interface UnsympRepository extends JpaRepository<Unsymp, Long> {
 	List<Unsymp> findByUser_Id(Long user_id);
 	Unsymp findByReview_IdAndUser_Id(Long review_id, Long user_id);
 }
